requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
sqlalchemy==2.0.15
pydantic==1.10.7
passlib==1.7.4
python-jose==3.3.0
python-multipart==0.0.6
boto3==1.26.142
python-dotenv==1.0.0